requirements.txt
streamlit
langchain
langchain-community
pypdf
langchain-text-splitters
langchain-cohere
faiss-cpu
google-generativeai
langchain-google-genai